Sending specific data by client
Client:
When the user transfers a file, user can decide which path he wants to save it to.
The validity of the entered paths is checked and added to the header.

The CreateHeader() function prepares specific headers for the string or file and sends the final data to the SendBytes() function to transmit it to the server.

Server:
Input of type String or file can now be sent in the SendBytes() function. (More optimized version of FileSend and StringSend functions)